Catch UnicodeErrors
Add catch for UnicodeError when executing command. Change .abproject file to be opened with "utf-8" encoding (using codecs), so even if DisplayName (or any other property) has value with non-ASCII symbols, it should be parsed correctly.
The only issue that remains is on Windows with Sublime 2 when path to project has non-ASCII symbols (for example cyrillic) - Python 2.7 cannot handle this case, so we are just showing error message.

Fixes http://teampulse.telerik.com/view#item/278007