Fix build and deploy when multiple devices
Fux Build and Deploy option when more than one devices are detected.
The code was trying to access 'name' property which is no longer provided by
`appbuilder device --json` command.
Also this code is failing when any of the properties is missing, for example
Genymotion does not have `model`.

Replace `name` with correct `displayName` and use `identifier` property as backup.
Use empty strings for backup of each property, so the code will work, no matter if it exists for the device.